feat: COLLATE, type affinity, last_insert_rowid, and LIKE ESCAPE
- F7.3d: COLLATE support in CompareValuesCollate (NOCASE, RTRIM, BINARY)
- F7.3d: Type affinity on INSERT via ApplyColumnAffinity
- F7.1c: LastInsertRowID() accessor on DB + engine tracking
- F7.3e: LIKE ESCAPE clause (BinaryOp.Escape field, escped recursive matching)
- F7.4c: Clear TODO in engine.go (index population noted)
- Fix test expectations for int64 vs float64 after affinity changes

+++ frigolite_query_test.go
@@ -448,7 +448,7 @@ func TestSelectAbs(t *testing.T) {
 	if res.Error != nil {
 		t.Fatalf("ABS: %v", res.Error)
 	}
-	if res.Rows[0][0] != float64(42) {
+	if res.Rows[0][0] != int64(42) {
 		t.Errorf("expected 42, got %v (type: %T)", res.Rows[0][0], res.Rows[0][0])
 	}
